refactor frontend structure
