+++ app/page.tsx
@@ -11,52 +11,33 @@ import {
   CircularProgress,
   Drawer,
   IconButton,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
   Paper,
   TextField,
   Toolbar,
   Typography,
   useMediaQuery,
   useTheme,
-  Divider,
   Dialog,
   DialogTitle,
   DialogContent,
   DialogContentText,
   DialogActions,
-  keyframes,
   Slide,
   Grid,
 } from "@mui/material"
 import {
   Menu as MenuIcon,
   Add,
-  PersonOutline,
-  History,
-  InfoOutlined,
   AutoAwesome,
   HelpOutline,
-  ErrorOutline,
 } from "@mui/icons-material"
 import ReactMarkdown from "react-markdown"
 import remarkGfm from "remark-gfm"
 import type { TransitionProps } from "@mui/material/transitions"
-
-/* ----------------------------- animations ----------------------------- */
-
-const fadeIn = keyframes`
-  from { opacity: 0; transform: translateY(10px); }
-  to   { opacity: 1; transform: translateY(0); }
-`
-
-const textFadeIn = keyframes`
-  from { opacity: 0; transform: translateY(5px); }
-  to   { opacity: 1; transform: translateY(0); }
-`
+import SidebarContent from "@/components/chat/SidebarContent"
+import ProcessingIndicator from "@/components/chat/ProcessingIndicator"
+import ErrorMessage from "@/components/chat/ErrorMessage"
+import { fadeIn } from "@/components/chat/animations"
 
 /* ------------- Slide transition that always keeps node mounted -------- */
 
@@ -167,261 +148,6 @@ export default function Chat() {
   const handleSuggestionClick = (prompt: string) => {
     setInput(prompt)
   }
-
-  /* ------------------- sidebar ------------------- */
-
-  const SidebarContent = () => (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        height: "100%",
-        p: 2,
-        bgcolor: "#2c3e50", // Darker, more modern sidebar
-        color: "white",
-      }}
-    >
-      {/* logo */}
-      <Box sx={{ display: "flex", alignItems: "center", gap: 1.5, mb: 2 }}>
-        <Box
-          sx={{
-            minWidth: 40,
-            height: 32,
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            bgcolor: "#D73027",
-            borderRadius: 2,
-            px: 1,
-            py: 0.5,
-          }}
-        >
-          <img
-            src="/images/ameritas-logo.png"
-            alt="ABACUS Logo"
-            style={{ maxWidth: "100%", maxHeight: "100%", objectFit: "contain" }}
-            onError={(e) => {
-              e.currentTarget.style.display = "none"
-              e.currentTarget.parentElement!.innerHTML =
-                '<span style="color:white;font-weight:bold;font-size:14px;">A</span>'
-            }}
-          />
-        </Box>
-        <Typography variant="h6" sx={{ fontWeight: "bold" }}>
-          ABACUS
-        </Typography>
-      </Box>
-
-      {/* new chat */}
-      <Button
-        variant="contained"
-        startIcon={<Add />}
-        onClick={() => {
-          setMessages([])
-          setInput("")
-        }}
-        sx={{
-          mb: 2,
-          bgcolor: "rgba(255 255 255 / .1)",
-          "&:hover": { bgcolor: "rgba(255 255 255 / .2)" },
-          justifyContent: "flex-start",
-        }}
-      >
-        New Chat
-      </Button>
-
-      <Divider sx={{ my: 1, borderColor: "rgba(255 255 255 / .15)" }} />
-
-      {/* history header */}
-      <Box sx={{ display: "flex", alignItems: "center", gap: 1, px: 2, py: 1 }}>
-        <History sx={{ color: "rgba(255,255,255,0.7)" }} />
-        <Typography variant="overline" color="rgba(255,255,255,0.7)">
-          History
-        </Typography>
-      </Box>
-
-      {/* history list */}
-      <Box sx={{ flexGrow: 1, overflow: "auto" }}>
-        {queryHistory.length ? (
-          <List dense>
-            {[...queryHistory].reverse().map((q, i) => (
-              <Box key={i} sx={{ animation: `${fadeIn} 0.5s ease ${i * 0.05}s both` }}>
-                <ListItem disablePadding>
-                  <ListItemButton
-                    onClick={() => {
-                      setInput(q)
-                      if (isMobile) setIsDrawerOpen(false)
-                    }}
-                    sx={{ borderRadius: 1 }}
-                  >
-                    <ListItemText
-                      primary={q}
-                      primaryTypographyProps={{
-                        noWrap: true,
-                        sx: { color: "#E0E0E0" },
-                      }}
-                    />
-                  </ListItemButton>
-                </ListItem>
-              </Box>
-            ))}
-          </List>
-        ) : (
-          <Typography sx={{ p: 2, textAlign: "center", color: "rgba(255,255,255,0.5)" }}>No history yet.</Typography>
-        )}
-      </Box>
-
-      {/* about & user */}
-      <Divider sx={{ my: 1, borderColor: "rgba(255 255 255 / .15)" }} />
-      <List dense>
-        <ListItemButton onClick={() => setIsAboutDialogOpen(true)} sx={{ borderRadius: 1 }}>
-          <ListItemIcon sx={{ minWidth: 40 }}>
-            <InfoOutlined sx={{ color: "rgba(255 255 255 / .7)" }} />
-          </ListItemIcon>
-          <ListItemText primary="About" />
-        </ListItemButton>
-      </List>
-
-      {/* mock user section */}
-      <Box
-        onClick={(e) => setUserMenuAnchorEl(e.currentTarget)}
-        sx={{
-          display: "flex",
-          alignItems: "center",
-          gap: 1.5,
-          p: 1,
-          mt: 1,
-          borderRadius: 1,
-          cursor: "pointer",
-          "&:hover": { bgcolor: "rgba(255 255 255 / .08)" },
-        }}
-      >
-        <Avatar sx={{ bgcolor: "#00529B" }}>
-          <PersonOutline />
-        </Avatar>
-        <Typography variant="body1">Koushik Sinha</Typography>
-      </Box>
-    </Box>
-  )
-
-  /* ------------------- processing indicator ------------------- */
-
-  const ProcessingIndicator = () => {
-    const statuses = ["Connecting to AI...", "Accessing Knowledge Base...", "Compiling response..."]
-    const [currentStatusIndex, setCurrentStatusIndex] = useState(0)
-    const delay = 1800 // Configurable delay in ms
-
-    useEffect(() => {
-      setCurrentStatusIndex(0) // Reset on mount
-
-      const interval = setInterval(() => {
-        setCurrentStatusIndex((prevIndex) => {
-          const nextIndex = prevIndex + 1
-          if (nextIndex === statuses.length) {
-            clearInterval(interval)
-            return prevIndex // Stay on the last status
-          }
-          return nextIndex
-        })
-      }, delay)
-
-      return () => clearInterval(interval) // Cleanup on unmount
-    }, [])
-
-    return (
-      <Box
-        sx={{
-          display: "flex",
-          gap: 2,
-          alignItems: "flex-start",
-          justifyContent: "flex-start",
-          animation: `${fadeIn} 0.5s ease both`,
-          maxWidth: "800px",
-          mx: "auto",
-          width: "100%",
-        }}
-      >
-        <Avatar
-          sx={{
-            bgcolor: "#D73027",
-            color: "white",
-            width: 40,
-            height: 40,
-            boxShadow: "3px 3px 7px #c8cacd, -3px -3px 7px #ffffff",
-          }}
-        >
-          <img
-            src="/images/ameritas-logo.png"
-            alt="ABACUS"
-            style={{
-              width: "70%",
-              height: "70%",
-              objectFit: "contain",
-            }}
-          />
-        </Avatar>
-        <Paper
-          sx={{
-            p: 2,
-            maxWidth: "80%",
-            bgcolor: "var(--background)",
-            color: "text.primary",
-            borderRadius: "20px 20px 20px 5px",
-            boxShadow: "5px 5px 10px #d9dbde, -5px -5px 10px #ffffff",
-            display: "flex",
-            alignItems: "center",
-          }}
-        >
-          <CircularProgress size={16} sx={{ mr: 1.5, color: "text.secondary" }} />
-          <Typography variant="body1" key={currentStatusIndex} sx={{ animation: `${textFadeIn} 0.5s ease-in-out` }}>
-            {statuses[currentStatusIndex]}
-          </Typography>
-        </Paper>
-      </Box>
-    )
-  }
-
-  /* ------------------- error message component ------------------- */
-  const ErrorMessage = ({ message }: { message: { id: string; content: string } }) => (
-    <Box
-      key={message.id}
-      sx={{
-        display: "flex",
-        gap: 2,
-        alignItems: "flex-start",
-        animation: `${fadeIn} 0.5s ease both`,
-        justifyContent: "flex-start",
-        maxWidth: "800px",
-        mx: "auto",
-        width: "100%",
-      }}
-    >
-      <Avatar
-        sx={{
-          bgcolor: theme.palette.error.main,
-          color: "white",
-          width: 40,
-          height: 40,
-          boxShadow: "3px 3px 7px #c8cacd, -3px -3px 7px #ffffff",
-        }}
-      >
-        <ErrorOutline />
-      </Avatar>
-      <Paper
-        sx={{
-          p: 2,
-          maxWidth: "80%",
-          bgcolor: "var(--background)",
-          color: "text.primary",
-          borderRadius: "20px 20px 20px 5px",
-          boxShadow: "5px 5px 10px #d9dbde, -5px -5px 10px #ffffff",
-          border: `1px solid ${theme.palette.error.light}`,
-        }}
-      >
-        <ReactMarkdown remarkPlugins={[remarkGfm]}>{message.content}</ReactMarkdown>
-      </Paper>
-    </Box>
-  )
 
   /* ==================================================================== */
