+++ components/chat/animations.ts
@@ -1,0 +1,12 @@
+"use client"
+import { keyframes } from "@mui/material"
+
+export const fadeIn = keyframes`
+  from { opacity: 0; transform: translateY(10px); }
+  to   { opacity: 1; transform: translateY(0); }
+`
+
+export const textFadeIn = keyframes`
+  from { opacity: 0; transform: translateY(5px); }
+  to   { opacity: 1; transform: translateY(0); }
+`
