+++ components/providers/theme-registry.tsx
